[feature] Graph setup and benchmark find (#8)
* add types to models

* set up ingress and write to inmemory janusgraph

* set up benchmark

* benchmark timing

* remove useless test

* more

* remove line

* try to insert debug line

* move debugger

* try this

* remove import

* add root?

* just where the fuck is this going?

* what in the good fuck

* this works on local [skip ci]

* try setting global resource path

* didn't work [skip ci]

* try to set up base directory

+++ server/app/models/DoseModel.scala
@@ -1,13 +1,15 @@
 package models
 
+import java.lang.Float
+
 object DoseModel {
 
   case class DoseId(value: Int) extends AnyVal {
     override def toString: String = value.toString
   }
 
-  case class DoseSize(value: BigDecimal) extends AnyVal {
-    override def toString: String = value.toString()
+  case class DoseSize(value: Float) extends AnyVal {
+    override def toString: String = value.toString
   }
 
   case class DoseUnit(value: String) extends AnyVal {
